requirements.txt
fastapi
uvicorn
python-multipart
pydantic
numpy
opencv-python
Pillow
scipy
scikit-image
prometheus-client
rich
redis
celery
sqlalchemy
psycopg2-binary
alembic
